refactor: update data types for sex, year_diagnosed, and day_of_week across database schema and entity models

+++ backend/src/main/java/com/uni/medicare/shared/entity/History.java
@@ -25,7 +25,7 @@ public class History {
     private String conditionStatus;
 
     @Column(name = "year_diagnosed")
-    private Short yearDiagnosed;
+    private Integer yearDiagnosed;
 
     @Column(nullable = false, updatable = false)
     private LocalDateTime recordedAt;
